requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.55.3
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
mysql-connector-python==9.1.0
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
tzdata==2024.2
zipp==3.21.0
